requirements.txt
openai
gradio
python-decouple
python-dotenv